Moving more resources from email service into appropriate services.
